moved HowToApplyPost911GiBill.jsx into components directory
